pyproject.toml
[project]
name = "budgeting-app"
version = "0.1.0"
description = ""
authors = [
    {name = "WilliamD"}
]
readme = "README.md"
requires-python = ">=3.10, <4"
dependencies = [
    "asgiref (>=3.8.1,<4.0.0)",
    "backports-tarfile (>=1.2.0,<2.0.0)",
    "build (>=1.2.2.post1,<2.0.0)",
    "cachecontrol (>=0.14.2,<0.15.0)",
    "cachetools (>=5.5.1,<6.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "cffi (>=1.17.1,<2.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "cleo (>=2.1.0,<3.0.0)",
    "crashtest (>=0.4.1,<0.5.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "distlib (>=0.3.9,<0.4.0)",
    "django (>=5.1.6,<6.0.0)",
    "dulwich (>=0.22.7,<0.23.0)",
    "fastjsonschema (>=2.21.1,<3.0.0)",
    "filelock (>=3.17.0,<4.0.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "google-api-core (>=2.24.1,<3.0.0)",
    "google-api-python-client (>=2.160.0,<3.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-cloud-core (>=2.4.1,<3.0.0)",
    "google-cloud-firestore (>=2.20.0,<3.0.0)",
    "google-cloud-storage (>=3.0.0,<4.0.0)",
    "google-crc32c (>=1.6.0,<2.0.0)",
    "google-resumable-media (>=2.7.2,<3.0.0)",
    "googleapis-common-protos (>=1.66.0,<2.0.0)",
    "graphene (>=3.4.3,<4.0.0)",
    "graphene-django (>=3.2.2,<4.0.0)",
    "graphql-core (>=3.2.6,<4.0.0)",
    "graphql-relay (>=3.2.0,<4.0.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "grpcio-status (>=1.70.0,<2.0.0)",
    "httplib2 (>=0.22.0,<0.23.0)",
    "idna (>=3.10,<4.0)",
    "importlib-metadata (>=8.6.1,<9.0.0)",
    "installer (>=0.7.0,<0.8.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.1.0,<5.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "more-itertools (>=10.6.0,<11.0.0)",
    "msgpack (>=1.1.0,<2.0.0)",
    "packaging (>=24.2,<25.0)",
    "pkginfo (>=1.12.0,<2.0.0)",
    "platformdirs (>=4.3.6,<5.0.0)",
    "poetry (>=2.0.1,<3.0.0)",
    "poetry-core (>=2.0.1,<3.0.0)",
    "promise (>=2.3,<3.0)",
    "proto-plus (>=1.26.0,<2.0.0)",
    "protobuf (>=5.29.3,<6.0.0)",
    "pyasn1 (>=0.6.1,<0.7.0)",
    "pyasn1-modules (>=0.4.1,<0.5.0)",
    "pycparser (>=2.22,<3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pyparsing (>=3.2.1,<4.0.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "rapidfuzz (>=3.12.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "rsa (>=4.9,<5.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "text-unidecode (>=1.3,<2.0)",
    "tomli (>=2.2.1,<3.0.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "trove-classifiers (>=2025.1.15.22,<2026.0.0.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "uritemplate (>=4.1.1,<5.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "virtualenv (>=20.29.1,<21.0.0)",
    "xattr (>=1.1.4,<2.0.0)",
    "zipp (>=3.21.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
